refactor: extract useKeyboardShortcuts hook from App.tsx
